requirements.txt
websockets
RPi.GPIO
gpiozero
evdev
python-dotenv
